Multiple changes to Errorhandling like catching errors from Connect-HPEILO and retried all different ways to start it

+++ src/ILO-Inventorizer/QueryILO.ps1
@@ -19,7 +19,9 @@ Function Get-DataFromILO {
 
             $findILO = Find-HPEiLO $srv;
             $iLOVersion = ([regex]"\d").Match($findILO.PN).Value;
-            $conn = Connect-HPEiLO -Address $srv -Username $login.Username -Password (ConvertFrom-SecureString -SecureString ($login.Password) -AsPlainText) -DisableCertificateAuthentication:($config.deactivateCertificateValidation);
+
+            $conn = Connect-HPEiLO -Address $srv -Username $login.Username -Password (ConvertFrom-SecureString -SecureString ($login.Password) -AsPlainText) -DisableCertificateAuthentication:($config.deactivateCertificateValidation) -ErrorAction Stop;
+
 
             Log 6 "`tQuerying PowerSupply" -IgnoreLogActive
             $powerSupply = ($conn | Get-HPEiLOPowerSupply);
@@ -215,7 +217,17 @@ Function Get-DataFromILO {
         Save-DataInCSV $report;
     }
     catch {
-        Save-Exception $_ ($_.Exception.Message.ToString());
+        $message = $_.Exception.Message.ToString();
+        if ($message -match "401") {
+            Save-Exception $_ "The ILO-Server $srv returned Unauthorized. Verify that the password in your login.json is correct and you are able to log into the ILO-Interface with it.";
+        }
+        elseif($message -match "SSL"){
+            Save-Exception $_ "The ILO-Server $srv returned an Error with SSL. Verify that your certificate is properly set or set the flag '-DeactivateCertificateValidationILO'.";    
+        }
+        else {
+            Save-Exception $_ ($_.Exception.Message.ToString());
+
+        }
     }
 }
 
@@ -238,10 +250,9 @@ Function Register-Directory {
     )
     try {
         if ((-not (Test-Path -Path $Path)) -and $ignoreError) {
-            New-Item -Path $Path -Force -ItemType Directory;
+            New-Item -Path $Path -Force -ItemType Directory | Out-Null;
         }
-        elseif (Test-Path -Path $path) {
-            
+        elseif (Test-Path -Path $path) {   
             $isDirectory = (Get-Item ($Path)) -is [System.IO.DirectoryInfo];
             if (-not $isDirectory) {
                 $Path = $Path | Split-Path -Parent -Resolve;
@@ -267,11 +278,17 @@ Function Save-DataInJSON {
     )
     try {
         $config = Get-Config;
-        Update-Config -ReportPath (Register-Directory ($config.reportPath)).ToString();
-        [string]$date = (Get-Date -Format "yyyy_MM_dd").ToString();
         $path = $config.reportPath
-        $name = "$path\ilo_report_$date.json";
-        $report | ConvertTo-Json -Depth 15 | Out-File -FilePath $name -Force;
+        if (Test-Path -Path ($path)) {
+
+            Update-Config -ReportPath (Register-Directory ($config.reportPath)).ToString();
+            [string]$date = (Get-Date -Format "yyyy_MM_dd").ToString();
+            $name = "$path\ilo_report_$date.json";
+            $report | ConvertTo-Json -Depth 15 | Out-File -FilePath $name -Force;
+        }
+        else {
+            throw [System.IO.DirectoryNotFoundException] "No appropriate path could be find to save the report files. Please verify the specified one exists. The current path is '$path'"
+        }
     }
     catch {
         Save-Exception $_ ($_.Exception.ToString());
@@ -292,140 +309,146 @@ Function Save-DataInCSV {
         [string]$date = (Get-Date -Format "yyyy_MM_dd").ToString();
         $path = $config.reportPath;
     
-        $name = "$path\ilo_report_$date.csv"
-        $inventoryData = Get-InventoryData;
+        if (Test-Path -Path $path) {
 
-        # General (like in Inventory)
-        if ((-not $config.ignoreMACAddress) -or (-not $config.ignoreSerialNumbers)) {
+            $name = "$path\ilo_report_$date.csv"
+            $inventoryData = Get-InventoryData;
 
-            $csv_report = @();
-            foreach ($sr in $Report) {
-                $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
-                $csv_report += [ordered]@{
-                    Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "-";
-                    Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
-                    Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
-                    Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
-                    MAC_1         = $sr.MAC_1.Length -gt 0 ? $sr.MAC_1: "-";
-                    MAC_2         = $sr.MAC_2.Length -gt 0 ? $sr.MAC_2 : "-";
-                    MAC_3         = $sr.MAC_3.Length -gt 0 ? $sr.MAC_3 : "-";
-                    MAC_4         = $sr.MAC_4.Length -gt 0 ? $sr.MAC_4 : "-";
-                    Mgnt_MAC      = $sr.Mgnt_MAC.Length -gt 0 ? $sr.Mgnt_MAC : "-";
-                }
-            }
-            $csv_report = Get-StandardizedCSV $csv_report;
-            $csv_report | ConvertTo-Csv -Delimiter ";" | Out-File -FilePath $name -Force;
-        }
-        
-        # MAC (if not deactivated)
-        if (-not $config.ignoreMACAddress) {
+            # General (like in Inventory)
+            if ((-not $config.ignoreMACAddress) -or (-not $config.ignoreSerialNumbers)) {
 
-            $name = "$path\ilo_report_MAC_$date.csv"
-            $csv_mac_report = @();
-            foreach ($sr in $Report) {
-                $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
-                $csv_mac = [ordered]@{
-                    Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "-";
-                    Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
-                    Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
-                    Mgnt_MAC      = $sr.Mgnt_MAC.Length -gt 0 ? $sr.Mgnt_MAC : "-";
-                }   
-
-                [int]$i = 1;
-                foreach ($nic in $sr.NetworkInterfaces) {
-                    $nic.MACAddress = $nic.MACAddress.Length -gt 0 ? $nic.MACAddress : "-";
-                    $csv_mac.Add(("NetInterf_MAC_$i"), $nic.MACAddress);
-                    $i++;
-                }
-
-                $i = 1
-                foreach ($nad in $sr.NetworkAdapter) {
-                    foreach ($p in $nad.Ports) {
-                        $p.MACAddress = $p.MACAddress.Length -gt 0 ? $p.MACAddress : "-";
-                        $csv_mac.Add(("NetAdap_MAC_$i"), $p.MACAddress);   
-                        $i++; 
+                $csv_report = @();
+                foreach ($sr in $Report) {
+                    $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
+                    $csv_report += [ordered]@{
+                        Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "-";
+                        Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
+                        Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
+                        Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
+                        MAC_1         = $sr.MAC_1.Length -gt 0 ? $sr.MAC_1: "-";
+                        MAC_2         = $sr.MAC_2.Length -gt 0 ? $sr.MAC_2 : "-";
+                        MAC_3         = $sr.MAC_3.Length -gt 0 ? $sr.MAC_3 : "-";
+                        MAC_4         = $sr.MAC_4.Length -gt 0 ? $sr.MAC_4 : "-";
+                        Mgnt_MAC      = $sr.Mgnt_MAC.Length -gt 0 ? $sr.Mgnt_MAC : "-";
                     }
                 }
-        
-                $csv_mac_report += $csv_mac
+                $csv_report = Get-StandardizedCSV $csv_report;
+                $csv_report | ConvertTo-Csv -Delimiter ";" | Out-File -FilePath $name -Force;
             }
-            $csv_mac_report = Get-StandardizedCSV $csv_mac_report;
-            $csv_mac_report | Export-Csv -Path $name -Delimiter ";" -Force;
-        }
-
-        # SerialNumber (if not deactivated)
-        if (-not $config.ignoreSerialNumbers) {
-            $name = "$path\ilo_report_SERIAL_$date.csv"
-            $csv_serial_report = @();
-            $csv_additional_info = @();
-            foreach ($sr in $Report) {
-                $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
-                $csv_serial = [ordered]@{
-                    Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "";
-                    Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
-                    Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
-                    Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
-                }   
-                $csv_additional = [ordered]@{
-                    Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "";
-                    Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
-                    Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
-                    Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
-                };
-
-                [int]$i = 1;
-                $pwr = $sr.PowerSupply.PowerSupplies;
-                foreach ($ps in $pwr) {
-                    $ps.Serial = $ps.Serial.Length -gt 0 ? $ps.Serial : "-";
-                    $csv_serial.Add(("PowerSupply$i"), $ps.Serial);
-                    $csv_additional.Add(("PowerSupply_$i"), $ps.Name);
-                    $i++;
-                }
         
-                $i = 1;
-                foreach ($pr in $sr.Processor) {
-                    $pr.Serial = $pr.Serial.Length -gt 0 ? $pr.Serial : "-";
-                    $csv_serial.Add(("Processor$i"), $pr.Serial);
-                    $csv_additional.Add(("Processor_$i"), $pr.Model);
-                    $i++;
-                }
+            # MAC (if not deactivated)
+            if (-not $config.ignoreMACAddress) {
 
-                $i = 1;
-                foreach ($dev in $sr.Devices) {
-                    if ($iLOVersion -gt 5) {
-                        $dev.Serial = $dev.Serial.Length -gt 0 ? $dev.Serial : "-";
-                        $csv_serial.Add(("Device$i"), $dev.Serial);
-                        $csv_additional.Add(("Device_$i"), $dev.Name);
+                $name = "$path\ilo_report_MAC_$date.csv"
+                $csv_mac_report = @();
+                foreach ($sr in $Report) {
+                    $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
+                    $csv_mac = [ordered]@{
+                        Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "-";
+                        Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
+                        Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
+                        Mgnt_MAC      = $sr.Mgnt_MAC.Length -gt 0 ? $sr.Mgnt_MAC : "-";
+                    }   
+
+                    [int]$i = 1;
+                    foreach ($nic in $sr.NetworkInterfaces) {
+                        $nic.MACAddress = $nic.MACAddress.Length -gt 0 ? $nic.MACAddress : "-";
+                        $csv_mac.Add(("NetInterf_MAC_$i"), $nic.MACAddress);
                         $i++;
                     }
-                }
 
-                $i = 1;
-                foreach ($stor in $sr.Storage) {
-                    if ($iLOVersion -lt 6) {
-                        $stor.Serial = $stor.Serial.Length -gt 0 ? $stor.Serial : "-";
-                        $csv_serial.Add(("Storage$i"), $stor.Serial);
-                        $csv_additional.Add(("Storage_$i"), $stor.Name + "," + $stor.Model);
+                    $i = 1
+                    foreach ($nad in $sr.NetworkAdapter) {
+                        foreach ($p in $nad.Ports) {
+                            $p.MACAddress = $p.MACAddress.Length -gt 0 ? $p.MACAddress : "-";
+                            $csv_mac.Add(("NetAdap_MAC_$i"), $p.MACAddress);   
+                            $i++; 
+                        }
+                    }
+        
+                    $csv_mac_report += $csv_mac
+                }
+                $csv_mac_report = Get-StandardizedCSV $csv_mac_report;
+                $csv_mac_report | Export-Csv -Path $name -Delimiter ";" -Force;
+            }
+
+            # SerialNumber (if not deactivated)
+            if (-not $config.ignoreSerialNumbers) {
+                $name = "$path\ilo_report_SERIAL_$date.csv"
+                $csv_serial_report = @();
+                $csv_additional_info = @();
+                foreach ($sr in $Report) {
+                    $inventorySrv = $inventoryData | Where-Object -Property "Hostname" -Contains -Value ($sr.Hostname);
+                    $csv_serial = [ordered]@{
+                        Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "";
+                        Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
+                        Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
+                        Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
+                    }   
+                    $csv_additional = [ordered]@{
+                        Label         = (($inventorySrv | Select-Object -Property "Label").Label).Length -gt 0 ? ($inventorySrv | Select-Object -Property "Label").Label : "";
+                        Hostname      = $sr.Hostname.Length -gt 0 ? $sr.Hostname : "-";
+                        Hostname_Mgnt = $sr.Hostname_Mgnt.Length -gt 0 ?  $sr.Hostname_Mgnt : "-";
+                        Serial        = $sr.Serial.Length -gt 0 ? $sr.Serial : "-";
+                    };
+
+                    [int]$i = 1;
+                    $pwr = $sr.PowerSupply.PowerSupplies;
+                    foreach ($ps in $pwr) {
+                        $ps.Serial = $ps.Serial.Length -gt 0 ? $ps.Serial : "-";
+                        $csv_serial.Add(("PowerSupply$i"), $ps.Serial);
+                        $csv_additional.Add(("PowerSupply_$i"), $ps.Name);
                         $i++;
                     }
-                }
+        
+                    $i = 1;
+                    foreach ($pr in $sr.Processor) {
+                        $pr.Serial = $pr.Serial.Length -gt 0 ? $pr.Serial : "-";
+                        $csv_serial.Add(("Processor$i"), $pr.Serial);
+                        $csv_additional.Add(("Processor_$i"), $pr.Model);
+                        $i++;
+                    }
 
-                $i = 1;
-                foreach ($mem in $sr.Memory) {
-                    $mem.Serial = $mem.Serial.Length -gt 0 ? $mem.Serial : "-";
-                    $csv_serial.Add(("Memory$i"), $mem.Serial)
-                    $csv_additional.Add(("Memory_$i"), $mem.Location)
-                    $i++;
+                    $i = 1;
+                    foreach ($dev in $sr.Devices) {
+                        if ($iLOVersion -gt 5) {
+                            $dev.Serial = $dev.Serial.Length -gt 0 ? $dev.Serial : "-";
+                            $csv_serial.Add(("Device$i"), $dev.Serial);
+                            $csv_additional.Add(("Device_$i"), $dev.Name);
+                            $i++;
+                        }
+                    }
+
+                    $i = 1;
+                    foreach ($stor in $sr.Storage) {
+                        if ($iLOVersion -lt 6) {
+                            $stor.Serial = $stor.Serial.Length -gt 0 ? $stor.Serial : "-";
+                            $csv_serial.Add(("Storage$i"), $stor.Serial);
+                            $csv_additional.Add(("Storage_$i"), $stor.Name + "," + $stor.Model);
+                            $i++;
+                        }
+                    }
+
+                    $i = 1;
+                    foreach ($mem in $sr.Memory) {
+                        $mem.Serial = $mem.Serial.Length -gt 0 ? $mem.Serial : "-";
+                        $csv_serial.Add(("Memory$i"), $mem.Serial)
+                        $csv_additional.Add(("Memory_$i"), $mem.Location)
+                        $i++;
+                    }
+                    $csv_additional_info += $csv_additional;
+                    $csv_serial_report += $csv_serial;
                 }
-                $csv_additional_info += $csv_additional;
-                $csv_serial_report += $csv_serial;
+                $csv_serial_report = Get-StandardizedCSV $csv_serial_report;
+                $csv_serial_report | Export-Csv -Path $name -Delimiter ";" -Force;
+                Add-Content -Path $name -Value "`r`n`n`n`nAdditional Information for above;"
+
+                $csv_additional_info = Get-StandardizedCSV $csv_additional_info; 
+                $csv_additional_info | ConvertTo-Csv -Delimiter ";" | Add-Content -Path $name -Force
             }
-            $csv_serial_report = Get-StandardizedCSV $csv_serial_report;
-            $csv_serial_report | Export-Csv -Path $name -Delimiter ";" -Force;
-            Add-Content -Path $name -Value "`r`n`n`n`nAdditional Information for above;"
-
-            $csv_additional_info = Get-StandardizedCSV $csv_additional_info; 
-            $csv_additional_info | ConvertTo-Csv -Delimiter ";" | Add-Content -Path $name -Force
+        }
+        else {
+            throw [System.IO.DirectoryNotFoundException] "No appropriate path could be find to save the report files. Please verify the specified one exists. The current path is '$path'"
         }
     }
     catch {
